fix: duplicate todos when creating, emtpy todo
