+++ src/App.css
@@ -37,7 +37,7 @@ input {
   border: none;
   outline: none;
   padding: 0 12px;
-  box-shadow: 2px 0 4px rgba(0,0,0, 0.2);
+  box-shadow: 0 0 4px rgba(0,0,0, 0.2);
   border-radius: 4px;
 }
 
@@ -45,11 +45,31 @@ input:focus {
   box-shadow: 1px 0 9px rgba(0,0,0, 0.25);
 }
 
+.Nav {
+  padding: 10px 25px;
+}
+
+.Nav__container {
+  display: flex;
+  align-items: center;
+}
+
+.Nav__image {
+  margin-right: 5px;
+}
+
+.Nav__text {
+  margin: 0;
+}
+
+.Todo__page {
+  width: 768px;
+}
+
 .ToDo__container {
   border: 1px solid rgba(0,0,0, 0.13);
   border-radius: 8px;
-  width: 500px;
-  margin-top: 5rem;
+  margin-top: 3rem;
   padding: 24px;
   box-shadow: 2px 2px 1px rgba(0,0,0, 0.09),
               3px 2px 3px rgba(0,0,0, 0.05);
@@ -57,16 +77,19 @@ input:focus {
 
 .Todo__creation {
   display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
 }
 
-.Todo__input {
-  flex: 1 1;
+.Todo__creation--text {
+  text-align: left;
+  font-size: 25px;
+  font-weight: 600;
 }
 
 .ToDo__list {
   display: flex;
   flex-direction: column;
-  margin-top: 1.5rem;
 }
 
 .ToDo__item {
@@ -109,7 +132,7 @@ input:focus {
 .Todo__toolbar {
   display: flex;
   justify-content: space-between;
-  margin-top: 24px;
+  margin-bottom: 24px;
 }
 
 .Todo__tabs {
